backend/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
